backend/microservices/user/requirements.txt
alembic
argon2_cffi
asgi_lifespan
asyncio
asyncpg
bcrypt
black
colorama
databases
docker
email-validator
fastapi
greenlet
httpx
isort
loguru
mypy
passlib
pathlib
pre-commit
psycopg2-binary
pydantic
pytest
pytest-asyncio
pytest-cov
pytest-xdist
python-decouple
python-decouple
python-dotenv
python-jose
python-slugify
redis
SQLAlchemy==2.0.0b3
sqlalchemy
trio
uvicorn
uvloop
